# api/docs/swagger.yaml
definitions:
...
  models.CreateQueue:
    properties:
      customer_id:
        type: string
      doctor_id:
        type: string
      queue_number:
        type: string
      queue_time:
        type: string
    type: object
  models.CreateSuperAdmin:
    properties:
      address:
        type: string
      author_id:
        type: string
      birth_date:
        type: string
      clinic_id:
        type: string
      drug_store_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
  models.Queue:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      deleted_at:
        type: string
      doctor_id:
        type: string
      id:
        type: string
      queue_number:
        type: string
      queue_time:
        type: string
      updated_at:
        type: string
    type: object
  models.QueuesResponse:
    properties:
      count:
        type: integer
      queues:
        items:
          $ref: '#/definitions/models.Queue'
        type: array
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.SuperAdmin:
    properties:
      address:
        type: string
      age:
        type: integer
      author_id:
        type: string
      birth_date:
        type: string
      clinic_id:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      drug_store_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  models.SuperAdminsResponse:
    properties:
      count:
        type: integer
      super_admins:
        items:
          $ref: '#/definitions/models.SuperAdmin'
        type: array
    type: object
...
  models.UpdateQueue:
    properties:
      customer_id:
        type: string
      doctor_id:
        type: string
      queue_number:
        type: string
      queue_time:
        type: string
    type: object
  models.UpdateSuperAdmin:
    properties:
      address:
        type: string
      author_id:
        type: string
      clinic_id:
        type: string
      drug_store_id:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
info:
  contact: {}
  description: Online doctor appointments and drug orders
  title: ShifoLink
  version: 1.0.0
paths:
...
  /queue:
    get:
      consumes:
      - application/json
      description: Get Queues list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QueuesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get Queues list
      tags:
      - queue
    post:
      consumes:
      - application/json
      description: Create a new Queue
      parameters:
      - description: Queue data
        in: body
        name: Queue
        required: true
        schema:
          $ref: '#/definitions/models.CreateQueue'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Queue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new Queue
      tags:
      - queue
  /queue/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Queue
      parameters:
      - description: Queue id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete Queue
      tags:
      - queue
    get:
      consumes:
      - application/json
      description: Get Queue by id
      parameters:
      - description: Queue
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Queue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get Queue by id
      tags:
      - queue
    put:
      consumes:
      - application/json
      description: Update Queue by id
      parameters:
      - description: Queue id
        in: path
        name: id
        required: true
        type: string
      - description: Queue
        in: body
        name: Queue
        required: true
        schema:
          $ref: '#/definitions/models.UpdateQueue'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Queue'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update Queue by id
      tags:
      - queue
  /super_admin:
    get:
      consumes:
      - application/json
      description: Get SuperAdmins list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuperAdminsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get SuperAdmins list
      tags:
      - super_admin
    post:
      consumes:
      - application/json
      description: Create a new SuperAdmin
      parameters:
      - description: SuperAdmin data
        in: body
        name: SuperAdmin
        required: true
        schema:
          $ref: '#/definitions/models.CreateSuperAdmin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SuperAdmin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new SuperAdmin
      tags:
      - super_admin
  /super_admin/{id}:
    delete:
      consumes:
      - application/json
      description: Delete SuperAdmin
      parameters:
      - description: SuperAdmin id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete SuperAdmin
      tags:
      - super_admin
    get:
      consumes:
      - application/json
      description: Get SuperAdmin by id
      parameters:
      - description: SuperAdmin
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuperAdmin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get SuperAdmin by id
      tags:
      - super_admin
    put:
      consumes:
      - application/json
      description: Update SuperAdmin by id
      parameters:
      - description: SuperAdmin id
        in: path
        name: id
        required: true
        type: string
      - description: SuperAdmin
        in: body
        name: SuperAdmin
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSuperAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuperAdmin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update SuperAdmin by id
      tags:
      - super_admin
swagger: "2.0"